[ONME-2510] Recalibrate software delay
delay_ns() loop was written on the assumption of one clock cycle per
iteration.

The loop GCC was compiling came out to 4 cycles per iteration on normal
builds, and 13 cycles per iteration in debug builds. The latter was slow
enough that we didn't switch to receive quickly enough after transmitting
to receive an ack, leading to ONME-2510 ("Connectivity issues with debug
build")

Convert the loop to assembler so we have a firmer idea of how long it will
take. Calibrate for a 3-cycle loop on M3/M4 - will be a bit slow on M0.
May be too fast on M7. (However, the set-up time itself is large enough
that we'll probably still take longer than the requested time for the
delays of up to 250ns used in the code.)

M4 calibration checked against the K64F's timers on a long delay.