Add mbed-cli configuration, fallback to NC
Add support for configuration via mbed-cli configuration mechanism,
and fall back to compiling with NC if we don't find any pin definitions.

ifdeffing out the code was causing link errors with the ARM linker - with
the NC define, all references to the radio driver should now be resolved
on initial link scan, after which an unused radio/6LoWPAN stack should go
on to be omitted.