Make inline assembler work with GCC+Thumb1
GCC insists on using a different assembler syntax for Thumb1. Adjust the
assembly code to cope.

Tweak the register directive to be "l" to make it a low register in GCC,
so it works for Thumb1. (Documentation says that's a synonym for "r" if
ARM).

IAR says that its "r" means a low register if Thumb, so not modifying that.