Disable power saving
Various reliability problems have been seen in sleepy nodes due to their
acks being missed, then them not being awake to ack a retransmission.

Some corrective action is being taken in the MAC to keep the driver on a
bit longer after sending an ack, to allow for retransmission, but it seems
that we are taking unnecessary risks by using all the power-saving
features of the chip - including the quiet transmission of ACKs if a
received data packets are loud.

As we've got lots of people demanding reliability, and no-one demanding
power saving for this driver, disable the power saving for now.

This does appear to improve big-ping-to-sleepy-node test reliability.