[ONME-2319] Enable RX safe mode
We have been seeing truncated packets without CRC errors, causing
reassembly errors.

Current hypothesis is that this may be due to the RX buffer being
overwritten by a following packet. Enabling RX safe mode eliminates
this race condition - after each packet is received, futher reception
is disabled until we've read the RX buffer.

Not yet proven that this was the cause of the problem, but enabling it
can't hurt.