Using mbed OS block SPI writes
Start to use block writes, to greatly improve transfer speed.

This opens up a new set of problems, in that a block write may well not
meet the spacing requirement of the Atmel radio which needs some spacing
between bytes. (250ns, according to the datasheet, but hard to pin down
from the diagrams where exactly this is measured - my assumption is it's
basically an extra 7.5/8MHz clock period).

As such, by default, halve the SPI frequency from 7.5MHz to 3.75MHz,
which meets the timing requirement if the block writes are totally
back-to-back, as they are on K64F for example.

Add JSON config to allow these speeds to be configured, and to allow
an assumption of spaced-enough SPI permitting 7.5MHz operation.

Pending discussion about a SPI spacing API for mbed OS, add code that
could use my proposal, switched on JSON.