Rework ack/promiscuous handling
When the "wait for ack" timer expired, we did a FORCE_PLL_ON, which
would break any ongoing reception. In a busy network, the following
sequence was quite common:

   * Send packet to node A
   * TRX_END, switch to promiscous mode, start ack timer
   * No ack arrives from A
   * Start receiving from node B
   * Ack timer expires, FORCE_PLL_ON, lose packet from B

Change to only manipulate the promiscuous bit on timer expiry, so as not
to disrupt other reception.

Also, shut off promiscuous and stop the timer when we receive something
other than the ack - the ack has to be immediate. This potentially
reduces a minor false-positive hole from using the same sequence number
as another node.

Also, improve reporting to Nanostack, so we tell it "TX_FAIL" when we
give up on an ack (due to timeout or seeing something else). This makes
us behave more like the simulator (which reports TX_FAIL when seeing
something else, but doesn't have a timer).